Fixed numerous bugs in txtParser. Moved input data to /data folder.
