--- health/src/main/testDomParser.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<data>
-	<column>[</column>
-    <column>,</column>
-    <column>p</column>
-	<column>d</column>
-	<column>c</column>
-	<column>]</column>
-</data>
